chore: rename Logo component
