+++ packages/libs/react-ui/src/components/BrandLogo/variants/Kadena.tsx
